backend/requirements.txt
fastapi
uvicorn[standard]
anthropic
openai
numpy
requests
python-dotenv
replicate
